Move colored output builder to libs/logger
